Fix transcripts article paths
Upon perusing the site, I discovered that the articles in the
`transcript` section failed to load. In the address bar, I noticed that
the url was something along the lines of `/transcripts///david-mccabe`.
Upon furher investigation, I realized that the `relativeDirectory`
property of the node was empty, thus resulting in this broken URL being
set as the path for the `transcript` article links.

This changes removes the `relativeDirectory` from the path, fixing the
links in the process.